Moved handler and initializer from example package
