+++ server/netty-server/src/main/scala/sttp/tapir/server/netty/NettyServerInitializer.scala
@@ -1,4 +1,4 @@
-package sttp.tapir.server.netty.example
+package sttp.tapir.server.netty
 
 import scala.concurrent.ExecutionContext
